- Gave creature and world their own class files independent of display   or pygame in general. - Moved display functionality to a separate display class - Small fix to prevent balls all moving to the right in the beginning
